Fix create metrics partitions dynamically for current month (#3)
Metrics insertion failed when partitions for node_metrics did not
exist for the target timestamp, resulting in errors such as:
`failed to insert node_metrics for node node-16e7dd9a-1 at 2025-07-17 06:00:00 +0000 UTC: ERROR: no partition of relation "node_metrics" found for row (SQLSTATE 23514)`

This change updates the partition creation logic to dynamically generate
partitions for the current month using the system date, rather than hardcoded
names. This matches the approach in `make generate-test-upload` and ensures
metrics can be ingested successfully regardless of when the test or upload is run.

Signed-off-by: Gloria Ciavarrini <gciavarrini@redhat.com>